feat: adicionar paginação avançada, filtros e estatísticas a todos os endpoints
